py-iot-sim/requirements.txt
numpy==2.3.0
paho-mqtt==2.1.0
